rpc gen - add object support
